mv __setup__
